morphony/pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "morphony"
version = "0.1.0"
description = "Autonomous research agent foundation scaffolding."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "aiosqlite>=0.20.0,<1.0.0",
  "httpx>=0.27.0,<1.0.0",
  "pydantic>=2.8.0,<3.0.0",
  "pydantic-settings>=2.3.0,<3.0.0",
  "pyyaml>=6.0.1,<7.0.0",
  "rich>=13.7.1,<14.0.0",
  "structlog>=24.2.0,<26.0.0",
  "typer>=0.12.3,<1.0.0",
]

[project.scripts]
agent = "morphony.cli:app"

[project.optional-dependencies]
dev = [
  "pyright>=1.1.390,<2.0.0",
  "pytest>=8.2.0,<9.0.0",
  "pytest-asyncio>=0.23.8,<1.0.0",
  "pytest-cov>=5.0.0,<7.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/morphony"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
addopts = "-q"

[tool.pyright]
include = ["src", "tests"]
pythonVersion = "3.12"
typeCheckingMode = "strict"
venvPath = "."
venv = ".venv"

